requirements.txt
auto_mix_prep==0.2.0
gunicorn==20.1.0
Werkzeug==2.2
googletrans
flask-cors==3.0.10
cloudpickle==1.6.0
Flask==2.0.1
pandas==1.4.1
requests==2.28.1
torch==2.0.1
joblib
Flask-Caching
faiss-cpu==1.7.4
transformers
flasgger==0.9.7b2

db-dtypes
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
google-cloud-core
google-cloud-storage
google-cloud-bigquery==3.3.0
google-cloud-videointelligence==2.5.1
google-crc32c
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0

python-dotenv==1.0.1
